+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ### 4.18.3.3
 
-- Tetris Add lines and level to High Score Dialog
 - Support Table data binding read_only_sort: true option to allow visual sorting without affecting model data
-- Pause on showing High Score Dialog
+- Tetris Add lines and level to High Score Dialog
+- Tetris Pause on showing High Score Dialog
+- Tetris Make High Scores -> Show a check menu item
+- Tetris Disable pause button upon showing High Score Dialog
 
 ### 4.18.3.2